Added 'datetime' to Mqtt Heartbeat Report
This change adds Date and Time ('datetime') to the
Mqtt Heartbeat Report, which can help diagnose the
Mqtt connection and/or client device problems.

Uptime is already reported, but without a datetime
context it is unknown if the retained Mqtt
informantion is current.

Note: Date and Time will not be reported if
NTP_SUPPORT is NOT enabled.

-