Add hook to allow from setting dst dates with julian day counts.  … If day of week = 7 or 8, the some on month + week = juliain day count if 7, 29 Feb is not counted, and the range is  1 - 365 if 8, 29 Feb is count, and the range is 0 - 365 This follows the posix formats described in: http://www.gnu.org/software/libc/manual/html_node/TZ-Variable.html
Jn
This specifies the Julian day, with n between 1 and 365. February 29 is never counted, even in leap years.

n
This specifies the Julian day, with n between 0 and 365. February 29 is counted in leap years.

Mm.w.d
This specifies day d of week w of month m. The day d must be between 0 (Sunday) and 6. The week w must be between 1 and 5; week 1 is the first week in which day d occurs, and week 5 specifies the last d day in the month. The month m should be between 1 and 12.