Use deveopment version of PlatformIO Core
We would be thankful if you could use a development version of PlatformIO Core. in 99.999% it's very stable but there sometimes the cases which we didn't expect, such as https://github.com/platformio/platformio-core/issues/3377

So, if you will see any issues with PlatformIO Core-dev, please report us and we will fix ASAP. Thanks!