relay: fix sync reentrancy lock
Fix not resetting the lock when checking the size.

Don't lock in the boot and in the processing, both are executed in CONT
context. Still lock the relaySync() itself, since it will call relayStatus().
This goes back to the original implementation, but does not seem to be
relevant right now.