sns: pzem004t initialization clean-up
Indexed address settings & build flags.
Old parsing code is still here, but depends on the old build flag presence.

Remove most of the singleton-related code, update terminal commands to
use ctx instead of just dumping debug strings.
Since the initialization happens in the sensor setup, and we know the
pzem pointer - simply pass the 'sensor' address to the function and
utilize the c++ local static variable feature that allows their use in
the function body and any lambdas declared within it.

Multiplexing is still happening with the index-math, due for a change at
a later date (...with serial / hwserial / stream changes...)