terminal: revert UnescapedText handling
Totally breaks on unicode input.

Main reason to add this in the first place was handling of
terminal escape sequences, but perhaps it would be better off to do
that on another layer.

For example, <Backspace> will inject an invisible symbol into the resulting
'chunk', at least one other thing to do is to validate after we append
things into an 'argv' that it at least looks like some kind of text.